move CoinMap to another package
